+++ Changelog.md
@@ -5,6 +5,7 @@ OpenCore Changelog
 - Added TimeMachine detection to picker
 - Added early preview version of BootLiquor
 - Fixed FS discovery on NVMe with legacy drivers
+- Added DirectGopCacheMode option for FB cache policy
 
 #### v0.5.6
 - Various improvements to builtin text renderer